ixgbe: remove residual fix about resetting big Tx queues
No need to keep residues of a fix which is replaced by another one.
This reverts commit 5a6d9897f91f6bb4b2
(residual fix about resetting big Tx queues).

+++ lib/librte_pmd_ixgbe/ixgbe_rxtx.c
@@ -1813,9 +1813,9 @@ ixgbe_reset_tx_queue(struct igb_tx_queue *txq)
 		volatile union ixgbe_adv_tx_desc *txd = &txq->tx_ring[i];
 		txd->wb.status = IXGBE_TXD_STAT_DD;
 		txe[i].mbuf = NULL;
-		txe[i].last_id = (uint16_t)i;
-		txe[prev].next_id = (uint16_t)i;
-		prev = (uint16_t)i;
+		txe[i].last_id = i;
+		txe[prev].next_id = i;
+		prev = i;
 	}
 
 	txq->tx_next_dd = (uint16_t)(txq->tx_rs_thresh - 1);
